+++ docs/conf.py
@@ -25,7 +25,7 @@ copyright = "2020, The django-voting-app authors"
 author = "The django-voting-app authors"
 
 # The full version, including alpha/beta/rc tags
-release = "1.2.1"
+release = "1.3.0"
 
 
 # -- General configuration ---------------------------------------------------