+++ src/django_voting_app/settings.py
@@ -116,6 +116,8 @@ LOGIN_URL = "login"
 LOGIN_REDIRECT_URL = "home"
 LOGOUT_REDIRECT_URL = "login"
 
+DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
+
 STATIC_ROOT = BASE_DIR / "staticfiles"
 MEDIA_ROOT = BASE_DIR / "media"
 STATICFILES_DIRS = [BASE_DIR / "statics"]